Refactor util
