+++ packages/core/src/common/number-util-test.ts
@@ -2,9 +2,11 @@ import * as assert from 'assert';
 import {describe as context, describe, it} from 'mocha';
 
 import {TestUtil} from '../misc/test-util';
+import {StepConstraint} from './constraint/step';
 import {
 	constrainRange,
 	getDecimalDigits,
+	getSuitableDraggingScale,
 	loopRange,
 	mapRange,
 } from './number-util';
@@ -143,6 +145,58 @@ describe('NumberUtil', () => {
 		context(`when ${JSON.stringify(testCase.args)}`, () => {
 			it(`should loop ${testCase.expected}`, () => {
 				assert.strictEqual(loopRange(...testCase.args), testCase.expected);
+			});
+		});
+	});
+});
+
+describe(getSuitableDraggingScale.name, () => {
+	[
+		{
+			params: {
+				constraint: undefined,
+				rawValue: -10,
+			},
+			expected: 1,
+		},
+		{
+			params: {
+				constraint: undefined,
+				rawValue: 1,
+			},
+			expected: 0.1,
+		},
+		{
+			params: {
+				constraint: undefined,
+				rawValue: 0.02,
+			},
+			expected: 0.001,
+		},
+		{
+			params: {
+				constraint: undefined,
+				rawValue: 0,
+			},
+			expected: 0.1,
+		},
+		{
+			params: {
+				constraint: new StepConstraint(1),
+				rawValue: 123,
+			},
+			expected: 0.1,
+		},
+	].forEach(({params, expected}) => {
+		context(`when ${JSON.stringify(params)}`, () => {
+			it('should estimate suitable scale', () => {
+				assert.ok(
+					TestUtil.closeTo(
+						getSuitableDraggingScale(params.constraint, params.rawValue),
+						expected,
+						DELTA,
+					),
+				);
 			});
 		});
 	});
