+++ packages/core/src/common/number-util.ts
@@ -1,3 +1,7 @@
+import {findConstraint} from './constraint/composite';
+import {Constraint} from './constraint/constraint';
+import {StepConstraint} from './constraint/step';
+
 export function mapRange(
 	value: number,
 	start1: number,
@@ -25,4 +29,42 @@ export function constrainRange(
 
 export function loopRange(value: number, max: number): number {
 	return ((value % max) + max) % max;
+}
+
+function findStep(constraint: Constraint<number> | undefined): number | null {
+	const c = constraint ? findConstraint(constraint, StepConstraint) : null;
+	if (!c) {
+		return null;
+	}
+
+	return c.step;
+}
+
+export function getSuitableDecimalDigits(
+	constraint: Constraint<number> | undefined,
+	rawValue: number,
+): number {
+	const sc = constraint && findConstraint(constraint, StepConstraint);
+	if (sc) {
+		return getDecimalDigits(sc.step);
+	}
+
+	return Math.max(getDecimalDigits(rawValue), 2);
+}
+
+export function getBaseStep(
+	constraint: Constraint<number> | undefined,
+): number {
+	const step = findStep(constraint);
+	return step ?? 1;
+}
+
+export function getSuitableDraggingScale(
+	constraint: Constraint<number> | undefined,
+	rawValue: number,
+): number {
+	const sc = constraint && findConstraint(constraint, StepConstraint);
+	const base = Math.abs(sc?.step ?? rawValue);
+
+	return base === 0 ? 0.1 : Math.pow(10, Math.floor(Math.log10(base)) - 1);
 }
