+++ packages/core/src/common/picker-util.ts
@@ -1,0 +1,8 @@
+import {PickerLayout} from './params';
+
+export function parsePickerLayout(value: unknown): PickerLayout | undefined {
+	if (value === 'inline' || value === 'popup') {
+		return value;
+	}
+	return undefined;
+}
